.ipynb_checkpoints/requirements-checkpoint.txt
langchain-google-genai
langchain-community
langchain-classic
langchain-text-splitters
pypdf
faiss-cpu
pytest
python-dotenv